# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Revise API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SourceInputSourceType = typeof SourceInputSourceType[keyof typeof SourceInputSourceType];


export const SourceInputSourceType = {
  youtube: 'youtube',
  text: 'text',
  pdf: 'pdf',
} as const;

export interface SourceInput {
  sourceType: SourceInputSourceType;
  /**
     * YouTube URL (required when sourceType is youtube)
     * @nullable
     */
  youtubeUrl?: string | null;
  /**
     * Title for the text/pdf source (required when sourceType is text or pdf)
     * @nullable
     */
  textTitle?: string | null;
  /**
     * Raw text content to generate study materials from (required when sourceType is text or pdf)
     * @nullable
     */
  textContent?: string | null;
  /**
     * Maximum number of flashcards to generate (5-100, default 10)
     * @nullable
     */
  maxFlashcards?: number | null;
}

export type SourceSourceType = typeof SourceSourceType[keyof typeof SourceSourceType];


export const SourceSourceType = {
  youtube: 'youtube',
  text: 'text',
  pdf: 'pdf',
} as const;

export type SourceStatus = typeof SourceStatus[keyof typeof SourceStatus];


export const SourceStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface Source {
  id: number;
  sourceType: SourceSourceType;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  videoId?: string | null;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  channelName?: string | null;
  status: SourceStatus;
  /** @nullable */
  errorMessage?: string | null;
  flashcardCount: number;
  knownCount: number;
  createdAt: string;
}

export type SourceDetailSourceType = typeof SourceDetailSourceType[keyof typeof SourceDetailSourceType];


export const SourceDetailSourceType = {
  youtube: 'youtube',
  text: 'text',
  pdf: 'pdf',
} as const;

export type SourceDetailStatus = typeof SourceDetailStatus[keyof typeof SourceDetailStatus];


export const SourceDetailStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface Flashcard {
  id: number;
  sourceId: number;
  question: string;
  answer: string;
  known: boolean;
  reviewCount: number;
  createdAt: string;
}

export interface PracticeQuestion {
  id: number;
  sourceId: number;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
  createdAt: string;
}

export interface SourceDetail {
  id: number;
  sourceType: SourceDetailSourceType;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  videoId?: string | null;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  channelName?: string | null;
  status: SourceDetailStatus;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  summary?: string | null;
  /**
     * JSON string containing the mind map tree structure
     * @nullable
     */
  mindMap?: string | null;
  createdAt: string;
  flashcardCount: number;
  knownCount: number;
  flashcards: Flashcard[];
  questions: PracticeQuestion[];
}

export interface FlashcardReview {
  known: boolean;
}

export interface Stats {
  totalSources: number;
  totalFlashcards: number;
  knownFlashcards: number;
  unknownFlashcards: number;
  sourcesProcessed: number;
}