# lib/api-zod/src/generated/types/sourceDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Revise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from './flashcard';
import type { PracticeQuestion } from './practiceQuestion';
import type { SourceDetailSourceType } from './sourceDetailSourceType';
import type { SourceDetailStatus } from './sourceDetailStatus';

export interface SourceDetail {
  id: number;
  sourceType: SourceDetailSourceType;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  videoId?: string | null;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  channelName?: string | null;
  status: SourceDetailStatus;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  summary?: string | null;
  /**
     * JSON string containing the mind map tree structure
     * @nullable
     */
  mindMap?: string | null;
  createdAt: string;
  flashcardCount: number;
  knownCount: number;
  flashcards: Flashcard[];
  questions: PracticeQuestion[];
}