# lib/api-zod/src/generated/types/sourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Revise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceInputSourceType } from './sourceInputSourceType';

export interface SourceInput {
  sourceType: SourceInputSourceType;
  /**
     * YouTube URL (required when sourceType is youtube)
     * @nullable
     */
  youtubeUrl?: string | null;
  /**
     * Title for the text/pdf source (required when sourceType is text or pdf)
     * @nullable
     */
  textTitle?: string | null;
  /**
     * Raw text content to generate study materials from (required when sourceType is text or pdf)
     * @nullable
     */
  textContent?: string | null;
  /**
     * Maximum number of flashcards to generate (5-100, default 10)
     * @nullable
     */
  maxFlashcards?: number | null;
}